+++ integrations/amazon_bedrock/src/haystack_integrations/components/retrievers/amazon_bedrock/__init__.py
@@ -1,0 +1,9 @@
+# SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
+#
+# SPDX-License-Identifier: Apache-2.0
+
+from haystack_integrations.components.retrievers.amazon_bedrock.knowledge_base_retriever import (
+    BedrockKnowledgeBaseRetriever,
+)
+
+__all__ = ["BedrockKnowledgeBaseRetriever"]